# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가시스템 공통기능 1차 API
  version: 0.1.0
servers:
- url: /api
security:
- sessionCookie: []
...
paths:
...
  /api/users:
    get:
      operationId: searchUsers
      summary: 사용자 검색
      x-related-requirements:
      - REQ-008
      - REQ-009
      - REQ-002
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-067
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: employeeNo
        in: query
        required: false
        schema:
          type: string
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: departmentCode
        in: query
        required: false
        schema:
          type: string
      - name: rankName
        in: query
        required: false
        schema:
          type: string
      - name: employmentStatus
        in: query
        required: false
        schema:
          type: string
      - name: roleCode
        in: query
        required: false
        schema:
          type: string
      - name: systemUseYn
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/users/{userId}/usage:
    patch:
      operationId: updateUserUsage
      summary: 사용여부 변경
      x-related-requirements:
      - REQ-010
      - REQ-004
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 사용여부 변경 요청자는 R09 write 권한을 가져야 한다
      - 사용여부 변경 입력은 Requirement Registry REQ-010, REQ-004, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 사용여부 변경 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 사용여부 변경 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 사용여부 변경 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 사용여부 변경의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 사용여부 변경 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserUsageRequest'
  /api/users/{userId}/roles:
    put:
      operationId: updateUserBusinessRoles
      summary: 사용자 업무 역할 변경
      x-related-requirements:
      - REQ-011
      - REQ-019
      - REQ-022
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      - REQ-068
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 사용자 업무 역할 변경 요청자는 R09 write 권한을 가져야 한다
      - 사용자 업무 역할 변경 입력은 Requirement Registry REQ-011, REQ-019, REQ-022의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 사용자 업무 역할 변경 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 사용자 업무 역할 변경 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 사용자 업무 역할 변경 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 사용자 업무 역할 변경의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 사용자 업무 역할 변경 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRolesRequest'
...
  /api/roles/{roleCode}:
    put:
      operationId: updateRole
      summary: 역할 기준 변경
      x-related-requirements:
      - REQ-016
      - REQ-017
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 역할 기준 변경 요청자는 R09 write 권한을 가져야 한다
      - 역할 기준 변경 입력은 Requirement Registry REQ-016, REQ-017, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 역할 기준 변경 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 역할 기준 변경 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 역할 기준 변경 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 역할 기준 변경의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 역할 기준 변경 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: roleCode
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
  /api/user-roles:
    get:
      operationId: listUserRoles
      summary: 사용자 역할 조회
      x-related-requirements:
      - REQ-018
      - REQ-023
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: employeeNo
        in: query
        required: false
        schema:
          type: string
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: roleCode
        in: query
        required: false
        schema:
          type: string
      - name: validOn
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: assignmentType
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: assignUserRole
      summary: 사용자 역할 부여
      x-related-requirements:
      - REQ-019
      - REQ-022
      - REQ-064
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      - REQ-068
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 사용자 역할 부여 요청자는 R09 write 권한을 가져야 한다
      - 사용자 역할 부여 입력은 Requirement Registry REQ-019, REQ-022, REQ-064, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 사용자 역할 부여 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-state-transitions:
      - not-assigned -> ACTIVE
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 사용자 역할 부여 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 사용자 역할 부여 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 사용자 역할 부여의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 사용자 역할 부여 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUserRoleRequest'
  /api/user-roles/{userRoleId}:
    patch:
      operationId: changeUserRole
      summary: 사용자 역할 변경
      x-related-requirements:
      - REQ-020
      - REQ-022
      - REQ-064
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      - REQ-068
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 사용자 역할 변경 요청자는 R09 write 권한을 가져야 한다
      - 사용자 역할 변경 입력은 Requirement Registry REQ-020, REQ-022, REQ-064, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 사용자 역할 변경 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 사용자 역할 변경 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 사용자 역할 변경 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 사용자 역할 변경의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 사용자 역할 변경 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: userRoleId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeUserRoleRequest'
    delete:
      operationId: revokeUserRole
      summary: 사용자 역할 회수
      x-related-requirements:
      - REQ-021
      - REQ-022
      - REQ-064
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      - REQ-068
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 사용자 역할 회수 요청자는 R09 write 권한을 가져야 한다
      - 사용자 역할 회수 입력은 Requirement Registry REQ-021, REQ-022, REQ-064, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 사용자 역할 회수 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-state-transitions:
      - ACTIVE -> REVOKED
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 사용자 역할 회수 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 사용자 역할 회수 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 사용자 역할 회수의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 사용자 역할 회수 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: userRoleId
        in: path
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KNUESESSION
...
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        message:
          type: string
          nullable: true
    ApiError:
      type: object
      required:
      - success
      - error
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            path:
              type: string
            timestamp:
              type: string
              format: date-time
    RoleCode:
      type: string
      enum:
      - R01
      - R02
      - R03
      - R04
      - R05
      - R06
      - R07
      - R08
      - R09
    UseYn:
      type: string
      enum:
      - Y
      - N
...
    AssignmentType:
      type: string
      enum:
      - POSITION_BASED
      - MANUAL
...
    UpdateUserUsageRequest:
      type: object
      required:
      - systemUseYn
      properties:
        systemUseYn:
          $ref: '#/components/schemas/UseYn'
    UpdateUserRolesRequest:
      type: object
      required:
      - roleCodes
      - assignmentType
      - approverUserId
      - effectiveStartDate
      properties:
        roleCodes:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
        assignmentType:
          $ref: '#/components/schemas/AssignmentType'
        approverUserId:
          type: string
        effectiveStartDate:
          type: string
          format: date
        effectiveEndDate:
          type: string
          format: date
          nullable: true
...
    UpdateRoleRequest:
      type: object
      required:
      - roleCode
      - roleName
      - purpose
      - useYn
      properties:
        roleCode:
          $ref: '#/components/schemas/RoleCode'
        roleName:
          type: string
        purpose:
          type: string
        grantCriteria:
          type: string
          nullable: true
        dataScopeDefault:
          type: string
          nullable: true
        useYn:
          $ref: '#/components/schemas/UseYn'
    AssignUserRoleRequest:
      type: object
      required:
      - userId
      - roleCode
      - assignmentType
      - approverUserId
      - effectiveStartDate
      properties:
        userId:
          type: string
        roleCode:
          $ref: '#/components/schemas/RoleCode'
        assignmentType:
          $ref: '#/components/schemas/AssignmentType'
        approverUserId:
          type: string
        effectiveStartDate:
          type: string
          format: date
        effectiveEndDate:
          type: string
          format: date
          nullable: true
    ChangeUserRoleRequest:
      type: object
      required:
      - roleCode
      - assignmentType
      - approverUserId
      - effectiveStartDate
      properties:
        roleCode:
          $ref: '#/components/schemas/RoleCode'
        assignmentType:
          $ref: '#/components/schemas/AssignmentType'
        approverUserId:
          type: string
        effectiveStartDate:
          type: string
          format: date
        effectiveEndDate:
          type: string
          format: date
          nullable: true
...